Migrate to Maven
- Add pom.xml
- Add .gitignore
- Move source files to /src/main/java/
