+++ src/main/java/ch/epfl/flamemaker/geometry2d/AffineTransformation.java
