+++ src/main/java/ch/epfl/flamemaker/geometry2d/Point.java
