+++ src/main/java/ch/epfl/flamemaker/gui/AffineTransformationsComponent.java
