+++ src/main/java/ch/epfl/flamemaker/gui/FlameBuilderPreviewComponent.java
