+++ src/main/java/ch/epfl/flamemaker/gui/FlameMaker.java
