+++ src/main/java/ch/epfl/flamemaker/gui/ObservableFlameBuilder.java
