+++ src/main/java/ch/epfl/flamemaker/ifs/IFSAccumulator.java
