+++ src/main/java/ch/epfl/flamemaker/ifs/IFSAccumulatorBuilder.java
