+++ src/main/java/ch/epfl/flamemaker/ifs/IFSMaker.java
